backend/requirements.txt
django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.5
django-cors-headers==3.14.0
daphne==4.0.0
channels==4.0.0
channels_redis
google-api-python-client==2.80.0
google-auth-oauthlib==1.0.0